+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-java changelog
 =====================
 
+[2022-11-10] Version 9.1.2
+--------------------------
+**Flex**
+- Adding two new authorization API 'user_roles' and 'gooddata' in version `v1`
+
+**Messaging**
+- Add new Campaign properties (MessageFlow, OptInMessage, OptInKeywords, OptOutMessage, OptOutKeywords, HelpMessage, HelpKeywords)
+
+**Twiml**
+- Add new speech models to `Gather`.
+
+
 [2022-11-02] Version 9.1.1
 --------------------------
 **Api**